cart progress, no checkout yet

+++ GameShop/GameShop-Backend/src/main/java/com/mcgill/ecse321/GameShop/dto/CartDto/CartResponseDto.java
@@ -50,7 +50,7 @@ public class CartResponseDto {
         return totalPrice;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(double totalPrice) {
         this.totalPrice = Math.round(totalPrice * 100.0) / 100.0;
     }
